di: Move LLVM type wrappers to a separate module
This way, the raw pointer fields marked as `pub(super)` are not
accessible outside the wrapper modules. This makes the modules
containing a logic safer and less error-prone.
