+++ src/llvm/mod.rs
@@ -1,6 +1,6 @@
 mod di;
-mod ir;
 mod iter;
+mod types;
 
 use std::{
     borrow::Cow,
